04-adk-samples/adk-sample-02-dborders-with-mcptoolbox-bigquery/agent/requirements.txt
google-adk==1.5.0
python-dotenv
toolbox-core 
langchain
google-genai
